refactor: 폴더 구조 변경
- 공통 컴포넌트를 common 폴더로 분리
- 메인 화면 컴포넌트를 views 폴더로 이동
- views 별 하위 컴포넌트 분리 (baseball, github)
