+++ week3/assignment/src/App.jsx
@@ -1,8 +1,8 @@
 // src/App.jsx
 import { useState } from 'react';
-import Header from './components/Header';
-import GitHubSearchPage from './components/GitHubSearch/GitHubSearchPage';
-import BaseballGamePage from './components/BaseballGame/BaseballGamePage';
+import Header from './components/common/Header';
+import GitHubSearch from './components/views/GitHubSearch';
+import BaseballGame from './components/views/BaseballGame';
 
 function App() {
   const [currentTab, setCurrentTab] = useState('github'); // 기본 탭은 github
@@ -12,9 +12,9 @@ function App() {
       <Header currentTab={currentTab} setCurrentTab={setCurrentTab} />
       <main className="container mx-auto p-4 max-w-xl">
         {currentTab === 'github' ? (
-          <GitHubSearchPage />
+          <GitHubSearch />
         ) : (
-          <BaseballGamePage />
+          <BaseballGame />
         )}
       </main>
     </div>
