+++ week3/assignment/src/components/common/Header.jsx
@@ -1,4 +1,3 @@
-// src/components/Header.jsx
 import React from 'react';
 
 function Header({ currentTab, setCurrentTab }) {
